+++ trunk/sdl/edlist.h
@@ -152,11 +152,11 @@ template <typename T>
     return entry;
   }
 
-  inline int GetLength()  const {  return Length;  }
+  inline size_t GetLength() const {  return Length;  }
 
-  inline bool IsEmpty()    const {  return (Length == 0);  }
+  inline bool IsEmpty() const {  return (Length == 0);  }
 
-  T& Get( size_t index )  {
+  T& Get(size_t index)  {
 #ifdef _OLX_DEBUG
     TIndexOutOfRangeException::ValidateRange(__OlxSourceInfo, index, 0, Length+1);
 #endif
